abstract 적용
