+++ src/main/java/com/example/object/callFee/NightlyDiscountPhone.java
@@ -6,41 +6,27 @@ import java.time.Duration;
 import java.util.ArrayList;
 import java.util.List;
 
-public class NightlyDiscountPhone {
+public class NightlyDiscountPhone extends AbstractPhone {
     private static final int LATE_NIGHT_HOUR = 22;
 
     private Money nightlyAmount;
     private Money regularAmount;
     private Duration seconds;
-    private List<Call> calls = new ArrayList<>();
-    private double taxRate;
 
     public NightlyDiscountPhone(Money nightlyAmount, Money regularAmount, Duration seconds, double taxRate) {
+        super(taxRate);
         this.nightlyAmount = nightlyAmount;
         this.regularAmount = regularAmount;
         this.seconds = seconds;
-        this.taxRate = taxRate;
     }
 
-    public void call(Call call) {
-        calls.add(call);
-    }
-
-    public List<Call> getCalls() {
-        return calls;
-    }
-
-    public Money calculateFee() {
-        Money result = Money.ZERO;
-
-        for (Call call : calls) {
-            if (call.getFrom().getHour() >= LATE_NIGHT_HOUR) {
-                result = result.plus(nightlyAmount.times(call.getDuration().getSeconds() / seconds.getSeconds()));
-            } else {
-                result = result.plus(regularAmount.times(call.getDuration().getSeconds() / seconds.getSeconds()));
-            }
+    @Override
+    protected Money calculateCallFee(Call call) {
+        if (call.getFrom().getHour() >= LATE_NIGHT_HOUR) {
+            return nightlyAmount.times(call.getDuration().getSeconds() / seconds.getSeconds());
         }
-
-        return result.plus(result.times(taxRate));
+        else {
+            return regularAmount.times(call.getDuration().getSeconds() / seconds.getSeconds());
+        }
     }
 }
